@@ -2,7 +2,7 @@ package com.nasgrad
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import com.nasgrad.api.model.IssueCategory
 import com.nasgrad.api.model.IssueType
 import com.nasgrad.utils.Helper
