@@ -1,6 +1,6 @@
 package com.nasgrad.api.model
 
-import android.support.annotation.Keep
+import androidx.annotation.Keep
 
 @Keep
 data class IssueCategory(
