@@ -1,12 +1,10 @@
 package com.nasgrad.issue
 
 import android.Manifest
-import android.content.Intent
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v4.content.ContextCompat
-import android.support.v7.app.AppCompatActivity
-import android.view.MenuItem
+import androidx.fragment.app.Fragment
+import androidx.core.content.ContextCompat
+import androidx.appcompat.app.AppCompatActivity
 import com.nasgrad.api.model.Issue
 import com.nasgrad.nasGradApp.R
 import com.nasgrad.utils.Helper
